askdasjd ©

+++ lib/rn/commands/books.rb
@@ -1,5 +1,5 @@
 require 'fileutils'
-require_relative '../paths'
+require_relative '../../helpers/paths'
 
 module RN
   module Commands
@@ -13,7 +13,7 @@ module RN
           def call(name:, **)
             if self.validate(name) then
               FileUtils.mkdir_p self.path(name)
-              warn "Cuaderno #{name} creado!!"
+              warn "Cuaderno '#{name.upcase}' creado!!"
             end
           end
 
@@ -29,13 +29,16 @@ module RN
 
           def call(name: nil, **options)
             if options[:global] then
-              FileUtils.rm_rf(Dir.glob("#{self.root}/*"))
-              warn "Acaba de eliminar todas las notas contenidas en el cuaderno global #{self.root}"
-            elsif !name.nil? then
+              if Dir.glob("#{self.root}/*").empty? then warn "No hay notas que eliminar"
+              else
+                FileUtils.rm_rf(Dir.glob("#{self.root}/*"))
+                warn "Acaba de eliminar todas las notas contenidas en el cuaderno global #{self.root}"
+              end
+            elsif !name.nil?
               FileUtils.remove_entry_secure self.path(name)
-              warn "Acaba de eliminar el cuaderno #{self.path(name)} y todas las notas allí contenidas"
+              warn "Acaba de eliminar el cuaderno '#{self.path(name).split("/")[-1].upcase}' y todas las notas allí contenidas"
+            else warn "No se realizó ninguna tarea ya que no ingresó ningún argumento. \n Acompañe el comando con '--global' para borrar las notas contenidas en el caderno global. \n O bien, con el nombre de un cuaderno existente que quiera eliminar junto con su contenido"
             end
-            warn "No se realizó ninguna tarea ya que no ingresó ningún argumento. \n Acompañe el comando con '--global' para borrar las notas contenidas en el caderno global. \n O bien, con el nombre de un cuaderno existente que quiera eliminar junto con su contenido"
           end
         end
 
@@ -44,8 +47,9 @@ module RN
           include Paths
 
           def call(*)
-            puts Dir.glob("#{self.root}/*")
+            puts Dir.glob("#{self.root}/*").map {|path| path.split("/")[-1]}
             # puts Dir.entries(self.root).select {|f| File.directory?(File.join(self.root,f))}
+            # puts Dir.glob("#{self.root}/*")
           end
         end
